code review fixes, renamed folder and file name,deleted commented codes, added comments to functions
