+++ artisan/runner/host/manifest.md
@@ -1,5 +1,5 @@
-# Package artisan.cmsee.stack/aps/host-runner:20220610125414219-9fc2475f4c Manifest
-*autogenerated using [Artisan CLI](https://github.com/gatblau/artisan) on 10 Jun 22 12:54 +0000*
+# Package artisan.cmsee.stack/aps/host-runner:20220613132212984-d723a97d94 Manifest
+*autogenerated using [Artisan CLI](https://github.com/gatblau/artisan) on 13 Jun 22 13:22 +0000*
 ## Function: install-standalone
 install standalone host-runner application.
 ### Variables:
